+++ src/renderervk/tr_world.c
@@ -247,7 +247,7 @@ static int R_DlightFace( srfSurfaceFace_t *face, int dlightBits ) {
 	int			i;
 	const dlight_t	*dl;
 
-	for ( i = 0; i < tr.refdef.num_dlights; i++ ) {
+	for ( i = 0; i < (int)tr.refdef.num_dlights; i++ ) {
 		if ( ! ( dlightBits & ( 1 << i ) ) ) {
 			continue;
 		}
@@ -272,7 +272,7 @@ static int R_DlightGrid( srfGridMesh_t *grid, int dlightBits ) {
 	int			i;
 	const dlight_t	*dl;
 
-	for ( i = 0 ; i < tr.refdef.num_dlights ; i++ ) {
+	for ( i = 0 ; i < (int)tr.refdef.num_dlights ; i++ ) {
 		if ( ! ( dlightBits & ( 1 << i ) ) ) {
 			continue;
 		}
@@ -305,7 +305,7 @@ static int R_DlightTrisurf( srfTriangles_t *surf, int dlightBits ) {
 	int			i;
 	const dlight_t	*dl;
 
-	for ( i = 0 ; i < tr.refdef.num_dlights ; i++ ) {
+	for ( i = 0 ; i < (int)tr.refdef.num_dlights ; i++ ) {
 		if ( ! ( dlightBits & ( 1 << i ) ) ) {
 			continue;
 		}
@@ -452,7 +452,7 @@ static void R_RecursiveLightNode( const mnode_t* node )
 		if ( node->visframe != tr.visCount )
 			return;
 
-		if ( node->contents != CONTENTS_NODE )
+		if ( (int)node->contents != (int)CONTENTS_NODE )
 			break;
 
 		children[0] = children[1] = qfalse;
@@ -550,7 +550,7 @@ void R_AddBrushModelSurfaces ( trRefEntity_t *ent ) {
 		
 		R_TransformDlights( tr.viewParms.num_dlights, tr.viewParms.dlights, &tr.or );
 
-		for ( i = 0; i < tr.viewParms.num_dlights; i++ ) {
+	for ( i = 0; i < (int)tr.viewParms.num_dlights; i++ ) {
 			dl = &tr.viewParms.dlights[i];
 			if ( !R_LightCullBounds( dl, bmodel->bounds[0], bmodel->bounds[1] ) ) {
 				tr.lightCount++;
@@ -647,7 +647,7 @@ static void R_RecursiveWorldNode( mnode_t *node, unsigned int planeBits, unsigne
 
 		}
 
-		if ( node->contents != CONTENTS_NODE ) {
+		if ( (int)node->contents != (int)CONTENTS_NODE ) {
 			break;
 		}
 
@@ -664,7 +664,7 @@ static void R_RecursiveWorldNode( mnode_t *node, unsigned int planeBits, unsigne
 		if ( dlightBits ) {
 			int	i;
 
-			for ( i = 0 ; i < tr.refdef.num_dlights ; i++ ) {
+			for ( i = 0 ; i < (int)tr.refdef.num_dlights ; i++ ) {
 				const dlight_t	*dl;
 				float		dist;
 
@@ -751,7 +751,7 @@ static mnode_t *R_PointInLeaf( const vec3_t p ) {
 
 	node = tr.world->nodes;
 	while( 1 ) {
-		if (node->contents != CONTENTS_NODE ) {
+		if ( (int)node->contents != (int)CONTENTS_NODE ) {
 			break;
 		}
 		plane = node->plane;
@@ -925,7 +925,7 @@ void R_AddWorldSurfaces( void ) {
 	// instead of having copypasted versions for both world and local cases
 
 	R_TransformDlights( tr.viewParms.num_dlights, tr.viewParms.dlights, &tr.viewParms.world );
-	for ( i = 0; i < tr.viewParms.num_dlights; i++ ) 
+	for ( i = 0; i < (int)tr.viewParms.num_dlights; i++ ) 
 	{
 		dl = &tr.viewParms.dlights[i];	
 		dl->head = dl->tail = NULL;
